requirements.txt
# Case Fictício - Teste Data Platform MVP -- Python Dependencies

# Data generation
faker>=22.0.0
pandas>=2.0.0

# GCP SDKs
google-cloud-storage>=2.14.0
google-cloud-bigquery>=3.17.0
google-cloud-functions-framework>=3.5.0

# Data processing
pyarrow>=14.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0
